moved event utility function into eventTypes and removed unused EventType
